[intel] Unconditionally reset PHY and restart autonegotiation
If the link is already up, we currently avoid performing a PHY reset,
since doing so on some models will drop the link speed to 10Mbps.
Debugging indicates that in these cases, the PHY has simply not
performed autonegotiation since the reset was issued.

Explicitly restart autonegotiation via the MII registers, allowing us
to always reset the PHY (and hence start in a more predictable state).

Signed-off-by: Michael Brown <mcb30@ipxe.org>